Replace all _or_404() by proper first() + flash() + redirect() and more security checks

+++ controllers/api.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, redirect, url_for, abort
+from flask import Blueprint, redirect, url_for, abort, flash
 from flask_security import login_required, current_user
 from models import db, Apitoken
 from utils import generate_uniques_apitoken
@@ -25,7 +25,11 @@ def apitoken_new():
 @bp_api.route('/apitoken/<string:apit>/del')
 @login_required
 def apitoken_del(apit):
-    apitoken = Apitoken.query.get_or_404(apit)
+    apitoken = Apitoken.query.filter(Apitoken.id == apit).first()
+    if not apitoken:
+        flash("API Token not found", "error")
+        return redirect(url_for("bp_main.home"))
+
     db.session.delete(apitoken)
     db.session.commit()
     return redirect(url_for('bp_users.user_profile'))
